requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.2.0
openai>=1.3.0
python-dotenv==1.0.1
supabase==2.3.5
redis==5.0.1
httpx==0.25.2
jinja2==3.1.2
